requirements.txt
httpx>=0.27.0
textual>=6.0.0
python-dotenv>=1.0.1
textual-image>=0.8.3
